Optimize Collections index performance Related to #567
Used a profiler to figure out which method causes the large delay of 7+
seconds when loading the page. It turns out that it's the part where
they instantiate Solr document objects for Solr repsone. Why it's so slow, who
knows, but I'm getting around it by using Solr response directly. This
means that I had to change all files involved in processing collection
index and the tests but the result is worth it. From my tests it seems
to speed up collection idex access 5-7 times.